Operators and Conditions

+++ First.java
@@ -5,7 +5,7 @@
         int num1= 12;
         int num2= 13;
         byte num3 = 14;
-       int num = num1 + num2 + num3;
+       int num = (byte)num1 + num2 + num3;
       System.out.println(num);
     }   
 }
